Force chunk load on elevation check
When `BlockState` is called on a block that is within a chunk that has yet to be
loaded, it will always return `nil`. This results in a situation wherein the height
found by the elevation check will always return -63Y in these cases.

This commmit attempts to fix this by forcing an asynchronous chunk load of the
destination chunk before elevation is calculated.

This is most apparent in ChunkyBorder, wherein a worldwrap teleport would attempt
to position the player in a chunk that had yet to be loaded, or even exist, always
teleporting them to -63Y

See: https://github.com/pop4959/ChunkyBorder/issues/91